Creates Tables 7.6 to 7.9

Table 7.6 Divorces by age of husband and age of wife

Table 7.7 Divorces by duration of marriage and age of husband

Table 7.8 Divorces by duration of marriage and age of wife

Table 7.9 Divorces by duration of marriage and number of dependent children

Usage

create_t7.6_to_7.9(
  data,
  group_var = "age_w",
  by_var = "age_h",
  data_year = NA,
  tablename = "Table_7_6",
  output_path = NULL
)

Arguments

data

data frame being used

group_var

the tables columns

by_var

tables rows

data_year

The year to report on. Defaults to the last in the data.

tablename

Name of the table to be saved as a csv file. Optional.

output_path

The path to export the generated csv table. Optional.

Value

data frame with tabulated results

Examples

t7.6 <- create_t7.6_to_7.9(div_data, data_year = 2019, group_var = "age_w", by_var = "age_h", tablename = "Table_7_6")
#> Error: object 'div_data' not found
t7.7 <- create_t7.6_to_7.9(div_data, data_year = 2019, group_var = "age_h", by_var = "dur_grp", tablename = "Table_7_7")
#> Error: object 'div_data' not found
t7.8 <- create_t7.6_to_7.9(div_data, data_year = 2019, group_var = "age_w", by_var = "dur_grp", tablename = "Table_7_8")
#> Error: object 'div_data' not found
t7.9 <- create_t7.6_to_7.9(div_data, data_year = 2019, group_var = "child", by_var = "dur_grp", tablename = "Table_7_9")
#> Error: object 'div_data' not found